split the create & set price task
